requirements.txt
python-dotenv
pyTelegramBotAPI
redis